setup sphinx documentation, rename module to sa
