+++ python/pycomposer/.gitignore
@@ -1,0 +1,5 @@
+sa-release/
+build/
+__pycache__/
+
+*.pyc
