+++ python/pycomposer/doc/source/api_reference.rst
@@ -1,0 +1,8 @@
+API Reference For Writing Annotations
+=====================================
+
+This page gives an overview of how to use the ``sa.annotation`` package for
+annotating functions.
+
+.. toctree::
+  sa.annotation
