+++ python/pycomposer/doc/source/getting_started.rst
@@ -1,0 +1,4 @@
+Getting Started with Split Annotations
+======================================
+
+Put a tutorial with code examples here.
