+++ python/pycomposer/doc/source/sa.annotated.numpy.rst
@@ -1,0 +1,2 @@
+``sa.annotated.numpy``
+======================
