+++ python/pycomposer/doc/source/sa.annotated.pandas.rst
@@ -1,0 +1,2 @@
+``sa.annotated.pandas``
+=======================
