+++ python/pycomposer/doc/source/sa.annotated.rst
@@ -1,0 +1,6 @@
+sa.annotated 
+=============
+
+.. toctree::
+    sa.annotated.numpy
+    sa.annotated.pandas
