+++ python/pycomposer/doc/source/sa.annotation.rst
@@ -1,0 +1,15 @@
+``sa.annotation``
+=================
+
+.. toctree::
+    sa.annotation.split_types
+
+.. automodule:: sa.annotation.sa
+    :members:
+    :special-members: __init__
+    :show-inheritance:
+
+.. automodule:: sa.annotation.annotation
+    :members: mut
+    :special-members: __init__
+    :show-inheritance:
