+++ python/pycomposer/doc/source/sa.annotation.split_types.rst
@@ -1,0 +1,10 @@
+``sa.annotation.split_types``
+=============================
+
+``sa.annotation.split_types.generics``
+--------------------------------------
+
+.. automodule:: sa.annotation.split_types.generics
+    :members:
+    :undoc-members:
+    :show-inheritance:
