+++ python/pycomposer/doc/source/sa.rst
@@ -1,0 +1,7 @@
+sa
+==
+
+.. toctree::
+    sa.annotation
+    sa.annotated
+
