--- python/pycomposer/pycomposer/__init__.py
@@ -1,7 +1,0 @@
-
-from .composer import sa, evaluate, mut
-from .split_types import SplitType, Broadcast
-from .vm.driver import STOP_ITERATION
-
-# Import the generics.
-from .split_types import A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S, T, U, V, W, X, Y, Z
