+++ python/pycomposer/sa/annotation/__init__.py
@@ -1,0 +1,2 @@
+from .sa import sa, evaluate, mut
+from .vm.driver import STOP_ITERATION
