+++ python/pycomposer/sa/annotation/sa.py
@@ -1,4 +1,3 @@
-
 from .annotation import Annotation, mut
 from .dag import LogicalPlan, evaluate_dag
 from .split_types import *
