+++ python/pycomposer/sa/annotation/split_types/split_types.py
@@ -212,31 +212,3 @@ class Broadcast(SplitType):
         return None
 
     def __str__(self): return "broadcast"
-
-# Convinience functions for creating named generics.
-A = lambda: GenericType("A")
-B = lambda: GenericType("B")
-C = lambda: GenericType("C")
-D = lambda: GenericType("D")
-E = lambda: GenericType("E")
-F = lambda: GenericType("F")
-G = lambda: GenericType("G")
-H = lambda: GenericType("H")
-I = lambda: GenericType("I")
-J = lambda: GenericType("J")
-K = lambda: GenericType("K")
-L = lambda: GenericType("L")
-M = lambda: GenericType("M")
-N = lambda: GenericType("N")
-O = lambda: GenericType("O")
-P = lambda: GenericType("P")
-Q = lambda: GenericType("Q")
-R = lambda: GenericType("R")
-S = lambda: GenericType("S")
-T = lambda: GenericType("T")
-U = lambda: GenericType("U")
-V = lambda: GenericType("V")
-W = lambda: GenericType("V")
-X = lambda: GenericType("X")
-Y = lambda: GenericType("Y")
-Z = lambda: GenericType("Z")
